+++ packages/core/global-navigation/docs/shared/index.js
@@ -2,6 +2,7 @@
 
 import React from 'react';
 import { Example } from '@atlaskit/docs';
+import { colors } from '@atlaskit/theme';
 
 /**
  * Load an example in an iframe
@@ -22,5 +23,20 @@ export const IframeExample = ({ source, title, url }: *) => (
     )}
     source={source}
     title={title}
+  />
+);
+
+/**
+ * Horizontal rule
+ */
+export const Hr = () => (
+  <hr
+    css={{
+      backgroundColor: colors.N40,
+      border: 0,
+      height: 2,
+      marginBottom: '3em',
+      marginTop: '3em',
+    }}
   />
 );