+++ packages/editor/editor-core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @atlaskit/editor-core
 
+## 82.2.5
+- [patch] Fix image markdown tip in help dialog [79465ec](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/79465ec)
+
+## 82.2.4
+- [patch] Add translations for help-dialog and colors [42a2916](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/42a2916)
+
+## 82.2.3
+- [patch] Add translation for quick insert [d6c438c](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/d6c438c)
+
+## 82.2.2
+- [patch] ED-5421: avoid adding PM history record for simple table interactions [c6cb409](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/c6cb409)
+
 ## 82.2.1
 - [patch] ED-5357: getting rid of ak components in table controls [83ca3e9](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/83ca3e9)